backend/requirements.txt
barcode_generator==0.1rc15
boto3==1.26.142
botocore==1.29.142
botocore==1.23.34
flasgger==0.9.7.1
Flask==2.0.3
Flask_Cors==3.0.10
img2pdf==0.4.4
MySQL-python==1.2.5
Pillow==9.5.0
Pillow==9.0.1
Pillow==9.5.0
pycryptodome==3.18.0
PyJWT==2.3.0
PyJWT==2.7.0
python-dotenv==1.0.0
qrcode==7.4.2
Werkzeug==2.0.3
